Simulate Spatial Error Process

This function generates a pure spatial error process, which is useful when you only want to simulate the error structure without including any deterministic part (i.e., no xb term). This can be used to analyze or simulate the behavior of spatially dependent errors in isolation.

Usage

sim_error(u, listw, lambda = 0.5, model = c("sar", "ma"))

Arguments

u

an error vector

listw

a listw object generated with sim_grid_listw().

lambda

a value value between -1 and 1. The spatial autoregressive coefficient for the error term.

model

default "sar". Which model should be simulated. Provide "ma" for the moving average.

Value

A numeric vector

References

See spreg.dgp.dgp_errproc

Examples

listw <- sim_grid_listw(5) 
u <- make_error(25)
sim_error(u, listw)
#>  [1]  0.72077234  1.21139365 -0.01628937  1.79836392  0.88627894  0.93390045
#>  [7] -0.22225632  1.35073306  0.42557045 -0.49278606  1.69043446  1.27781450
#> [13]  1.18502044  0.60722753 -0.95161448  1.28886302 -0.30478446 -0.17557889
#> [19] -0.25579186  0.07592074 -0.30797951 -1.34477749 -0.20967252  0.38866564
#> [25]  0.23639659